Refactor: Extract ReciterBox to a separate file
